requirements.txt
discord.py
pymongo
dnspython
python-dotenv
certifi>=2023.7.22
